Use the bus channel as storage for current state
Bounded bus-queue channels are ring buffers, so "overflowing" the channel just drops old values.
Keeping a subscriber that is never read from (only cloned to give to new widgets) means the
new widgets would always receive the last message (== current state, when msgs are state snapshots)
via the subscription, so we don't have to have a shared location for the state really.